requirements.txt
# Core Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.1,<1.0.0
pydantic>=2.7.1,<3.0.0
pydantic-settings>=2.2.0,<3.0.0

# Frontend
streamlit>=1.37.0,<2.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector Database
chromadb>=1.5.5,<2.0.0

# LLM and HTTP
openai>=1.35.3,<2.0.0
httpx>=0.27.0,<1.0.0
aiohttp>=3.9.5,<4.0.0
requests>=2.31.0,<3.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
pyyaml>=6.0.1,<7.0.0

# Data Processing
pandas>=2.2.2,<3.0.0
numpy>=1.26.4,<3.0.0

# Testing
pytest>=8.2.2,<9.0.0
pytest-asyncio>=0.23.7,<1.0.0

# Development
black>=24.4.2,<25.0.0
isort>=5.13.2,<6.0.0
mypy>=1.10.0,<2.0.0
